refactor(accounts): rename service and model files
