+++ pkg/accounts/mod.ts
@@ -30,14 +30,14 @@ import { AuthenticationService } from './service/authenticate.js';
 import { EditAccountService } from './service/editAccount.js';
 import { EtagVerifyService } from './service/etagGenerateVerify.js';
 import { FetchAccountService } from './service/fetchAccount.js';
-import { FollowService } from './service/follow.js';
-import { FreezeService } from './service/freeze.js';
-import { RegisterAccountService } from './service/register.js';
-import { ResendVerifyTokenService } from './service/resendToken.js';
+import { FollowService } from './service/followAccount.js';
+import { FreezeService } from './service/freezeAccount.js';
+import { RegisterAccountService } from './service/registerAccount.js';
+import { ResendVerifyTokenService } from './service/resendAccountVerifyToken.js';
 import { DummySendNotificationService } from './service/sendNotification.js';
-import { SilenceService } from './service/silence.js';
+import { SilenceService } from './service/silenceAccount.js';
 import { TokenGenerator } from './service/tokenGenerator.js';
-import { UnfollowService } from './service/unfollow.js';
+import { UnfollowService } from './service/unfollowAccount.js';
 
 export const accounts = new OpenAPIHono();
 const accountRepository = new InMemoryAccountRepository();
