+++ pkg/accounts/model/followAccount.test.ts
@@ -3,7 +3,10 @@ import { describe, expect, it } from 'vitest';
 
 import type { ID } from '../../id/type.js';
 import type { AccountID } from './account.js';
-import { AccountFollow, type CreateAccountFollowArgs } from './follow.js';
+import {
+  AccountFollow,
+  type CreateAccountFollowArgs,
+} from './followAccount.js';
 
 describe('AccountFollow', () => {
   it('generate new instance', () => {
