+++ pkg/accounts/model/followAccount.ts
