+++ pkg/accounts/model/repository.ts
@@ -3,7 +3,7 @@ import type { Option, Result } from '@mikuroxina/mini-fn';
 import { type ID } from '../../id/type.js';
 import type { Account } from './account.js';
 import { type AccountID } from './account.js';
-import type { AccountFollow } from './follow.js';
+import type { AccountFollow } from './followAccount.js';
 import type { InactiveAccount } from './inactiveAccount.js';
 
 export interface AccountRepository {
