+++ pkg/accounts/service/authenticateAccount.test.ts
@@ -5,7 +5,7 @@ import type { ID } from '../../id/type.js';
 import { Argon2idPasswordEncoder } from '../../password/mod.js';
 import { InMemoryAccountRepository } from '../adaptor/repository/dummy.js';
 import { Account } from '../model/account.js';
-import { AuthenticationService } from './authenticate.js';
+import { AuthenticationService } from './authenticateAccount.js';
 import { TokenGenerator } from './tokenGenerator.js';
 
 describe('AuthenticationService', () => {
