+++ pkg/accounts/service/authenticateAccount.ts
