+++ pkg/accounts/service/freezeAccount.ts
