+++ pkg/accounts/service/registerAccount.ts
