+++ pkg/notes/mod.ts
@@ -12,14 +12,14 @@ import { AuthenticationService } from '../accounts/service/authenticate.js';
 import { EditAccountService } from '../accounts/service/editAccount.js';
 import { EtagVerifyService } from '../accounts/service/etagGenerateVerify.js';
 import { FetchAccountService } from '../accounts/service/fetchAccount.js';
-import { FollowService } from '../accounts/service/follow.js';
-import { FreezeService } from '../accounts/service/freeze.js';
-import { RegisterAccountService } from '../accounts/service/register.js';
-import { ResendVerifyTokenService } from '../accounts/service/resendToken.js';
+import { FollowService } from '../accounts/service/followAccount.js';
+import { FreezeService } from '../accounts/service/freezeAccount.js';
+import { RegisterAccountService } from '../accounts/service/registerAccount.js';
+import { ResendVerifyTokenService } from '../accounts/service/resendAccountVerifyToken.js';
 import { DummySendNotificationService } from '../accounts/service/sendNotification.js';
-import { SilenceService } from '../accounts/service/silence.js';
+import { SilenceService } from '../accounts/service/silenceAccount.js';
 import { TokenGenerator } from '../accounts/service/tokenGenerator.js';
-import { UnfollowService } from '../accounts/service/unfollow.js';
+import { UnfollowService } from '../accounts/service/unfollowAccount.js';
 import { SnowflakeIDGenerator } from '../id/mod.js';
 import { AccountModule } from '../intermodule/account.js';
 import { Argon2idPasswordEncoder } from '../password/mod.js';
